requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-talisman==0.8.1
pymssql==2.2.5
asterisk-ami==0.1.7
google-cloud-speech==2.11.0
google-cloud-storage==2.1.0
requests==2.26.0
python-dateutil==2.8.2
apscheduler==3.9.1
pjsua==2.13
PyJWT==2.3.0
redis==4.3.4
cryptography==37.0.4
bcrypt==3.2.2
# Test dependencies
pytest==7.3.1
pytest-mock==3.10.0
mock==5.0.0
coverage==7.2.7